requirements.txt
pillow
numpy
opencv-python
matplotlib
scipy
click